requirements.txt
langchain
langgraph
langchain_community
langchain_classic
langchain_core
streamlit
langchain_groq
langchain_huggingface
pypdf
arxiv
wikipedia
ddgs
faiss-cpu
python-dotenv
ipykernel
sentence-transformers
validators
unstructured
youtube_search
youtube-transcript-api
pytube
langchain-tavily